Rename .h files to .hpp for sonar to mark them as C++
