+++ CMakeLists.txt
@@ -12,7 +12,7 @@ conan_basic_setup(TARGETS)
 add_library(CPP_MCTS INTERFACE)
 target_compile_features(CPP_MCTS INTERFACE cxx_override cxx_auto_type cxx_constexpr cxx_range_for)
 target_include_directories(CPP_MCTS INTERFACE include)
-set_target_properties(CPP_MCTS PROPERTIES PUBLIC_HEADER "include/mcts/mcts.h;include/mcts/graphviz.h")
+set_target_properties(CPP_MCTS PROPERTIES PUBLIC_HEADER "include/mcts/mcts.hpp;include/mcts/graphviz.hpp")
 install(TARGETS CPP_MCTS PUBLIC_HEADER DESTINATION include/mcts)
 
 if (CPP_MCTS_BUILD_SAMPLES)
