+++ samples/tictactoe/Game.cpp
@@ -1,7 +1,7 @@
 
 #include <QApplication>
 
-#include "GUI.h"
+#include "GUI.hpp"
 
 int main(int argc, char **argv){
 	QApplication app (argc, argv);
